Do not redeploy on bintray util-interface
util-interface is not versioned; if we deploy it from both the
2.12 and the 2.11 build to bintray, on the second deploy we
receive a "409" error, since the files are already present.

A better alternative would be implementing the "override"
bintray feature in dbuild by adding "?override=1" to the
upload uri. In the meantime, we just exclude "util-interface"
from the 2.11 bintray deploy section.