Move repos to proxy-ch
proxy-ch.typesafe.com used to have a configuration that was
frozen, due to a set of repository layouts once set when an
older version of Artifactory had a license key, and which
became unchangeable when it was upgraded to a newer, but OSS
version.

The configuration of Artifactory is now finally fixed, so
the list of remote repos can be updated there, and there is
no longer any need for custom repos in this build file.